% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{HD2013SGIdata}
\alias{HD2013SGIdata}
\title{Download a dataset from the HD2013SGI package}
\usage{
HD2013SGIdata(
  dataset = c("datamatrix", "datamatrixfull", "featuresPerWell", "Interactions",
    "mainEffects", "nrOfInteractionsPerTarget", "QueryAnnotation", "stabilitySelection",
    "TargetAnnotation")
)
}
\arguments{
\item{dataset}{The name of the dataset to download. This should be one of the
following: "datamatrix", "datamatrixfull", "featuresPerWell", "Interactions",
"mainEffects", "nrOfInteractionsPerTarget", "QueryAnnotation",
"stabilitySelection", or "TargetAnnotation".}
}
\value{
The requested dataset
}
\description{
This function downloads a dataset from the HD2013SGI package and caches it
locally. If the dataset has already been downloaded, it will be loaded from
the cache instead of downloading it again.
}
\details{
It replaces the previous datasets that were included in the package.
}
\examples{
\donttest{
HD2013SGIdata("datamatrix")
}
}
