\name{QueryAnnotation}
\alias{QueryAnnotation}
\docType{data}
\title{Annotation of all query genes in the screen}
\description{Annotation of all query genes in the screen.}
\details{This dataset is no longer directly included in this package. Instead, use the \code{HD2013SGIdata} function to download it.}
\usage{data(QueryAnnotation)}
\format{
  A data frame with 20 observations on the following 2 variables.
  \describe{
    \item{\code{ID}}{a character vector}
    \item{\code{Symbol}}{a character vector}
  }
}
\references{Laufer, Fischer et al., 2013}
\seealso{
\code{\link{HD2013SGI}}
}
\keyword{datasets}
